Add files via upload
In the first screenshot (1.png) it shows the interface of the game in which players can move the slider which changes the winning probability and the amount multiplier. Suppose if you play with a winning rate of 50% and put Rs.10 on the line, if you win you get Rs 19 and if you lose you lose Rs.10 which clearly shows the house edge of 1% average. Players often use strategies like Fibonacci, Martingale which are of no use in the long term. Martingale strategy is a strategy in which if you start with Rs.1 and win you will keep on betting with Rs.1 until you lose.....and if you lose you will double the bet amount like Rs.1 to Rs.2 and if you again lose doubling it to Rs.4 until you win to recover your previous losses.The Martingale strategy is one of the oldest betting systems in gambling. It’s based on a simple principle: Double your bet after every loss so that when you finally win, you recover all previous losses and make a profit equal to your original stake. In my basic code(Using C Programming) (2.png) I created a random number guessing game like dice and gave it a condition to run in loop untill the player losses 20 times in a row using the Martingale strategy . The basic algorithms of stake and other platforms also works in a similiar way where the house gets the profits in the longer run. Even if a player think that if he/she increases the winning percentage upto 70%, lowering the multiplier upto 1.4 times the money he/she is betting, it still wont work as the algorithm works in such a way that even if the multiplier decreases from 1.9x to 1.4x increasing your pseudo chances of winning it still wont work in the longer run as shown in (3.png) and (4.png). And in the last page (5.png) I gave a rough estimations of how much a player can lose if he is playing for Rs.1 profits every bet. If the player starts using Martingale Strategy with Rs.1, after 20 losses in a row he will lose Rs. 5,24,288 just trying to recover his Rs.1 loss. And if the player thinks that he/she will play with more winning chances of 70% then also will end up losing Rs. 16,384 trying to recover the previous losses. Therefore, this simple project shows that there is no easy way to earn money. Even with the perfect execution of Martingale Strategy players will still lose in the long term as mathematics and the algorithms always favours the house."Gambling doesn’t just drain money, it drains families and futures."